% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{authorities}
\alias{authorities}
\title{VIAF Authorities}
\format{
A tibble with 55 rows and 2 variables.
}
\usage{
data(authorities)
}
\description{
A dataset containing the names and schemes of 55 organizations
and libraries that participate in the VIAF.
}
\keyword{datasets}
