% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterdata.R
\name{cluster_data}
\alias{cluster_data}
\alias{cluster_summary}
\alias{cluster_sources}
\alias{cluster_authorship}
\alias{cluster_covers}
\alias{cluster_ISBNs}
\alias{cluster_jointcorps}
\alias{cluster_jointpersons}
\alias{cluster_links}
\alias{cluster_names}
\alias{cluster_titles}
\title{Get data tibbles from VIAF cluster}
\usage{
cluster_data(
  viafids = NULL,
  xslt = NULL,
  t = NULL,
  cache = tempdir(),
  intercache = tempdir(),
  redirectlog = file.path(tempdir(), "redirects.csv"),
  ...
)

cluster_summary(
  viafids = NULL,
  t = NULL,
  cache = tempdir(),
  intercache = tempdir(),
  redirectlog = file.path(tempdir(), "redirects.csv"),
  ...
)

cluster_sources(
  viafids = NULL,
  t = NULL,
  cache = tempdir(),
  intercache = tempdir(),
  redirectlog = file.path(tempdir(), "redirects.csv"),
  ...
)

cluster_authorship(
  viafids = NULL,
  t = NULL,
  cache = tempdir(),
  intercache = tempdir(),
  redirectlog = file.path(tempdir(), "redirects.csv"),
  ...
)

cluster_covers(
  viafids = NULL,
  t = NULL,
  cache = tempdir(),
  intercache = tempdir(),
  redirectlog = file.path(tempdir(), "redirects.csv"),
  ...
)

cluster_ISBNs(
  viafids = NULL,
  t = NULL,
  cache = tempdir(),
  intercache = tempdir(),
  redirectlog = file.path(tempdir(), "redirects.csv"),
  ...
)

cluster_jointcorps(
  viafids = NULL,
  t = NULL,
  cache = tempdir(),
  intercache = tempdir(),
  redirectlog = file.path(tempdir(), "redirects.csv"),
  ...
)

cluster_jointpersons(
  viafids = NULL,
  t = NULL,
  cache = tempdir(),
  intercache = tempdir(),
  redirectlog = file.path(tempdir(), "redirects.csv"),
  ...
)

cluster_links(
  viafids = NULL,
  t = NULL,
  cache = tempdir(),
  intercache = tempdir(),
  redirectlog = file.path(tempdir(), "redirects.csv"),
  ...
)

cluster_names(
  viafids = NULL,
  t = NULL,
  cache = tempdir(),
  intercache = tempdir(),
  redirectlog = file.path(tempdir(), "redirects.csv"),
  ...
)

cluster_titles(
  viafids = NULL,
  t = NULL,
  cache = tempdir(),
  intercache = tempdir(),
  redirectlog = file.path(tempdir(), "redirects.csv"),
  ...
)
}
\arguments{
\item{viafids}{A list of VIAF IDs to be retrieved}

\item{xslt}{The xslt stylesheet in character format.}

\item{t}{A tibble that contains summaries. If the columns of t are
identical to the columns retrieved in this function, the retrieved
rows are added to t. If not, a warning is given and t returned without
new rows. If t is NULL, a new tibble is returned with the columns
from the transformed viaf.xml.}

\item{cache}{A folder for storing download viaf.xml files and avoiding
unnecessary internet calls. The default is tempdir(), which is not
permanent. Researchers may want to substute a permanent folder to
maintain a snapshot of their data.}

\item{intercache}{A folder for storing the intermediate xml files produced
by the xslt transform. The intermediary files}

\item{redirectlog}{Path to csv file. The viaf api redirects superceded ids.
This parameter provides for logging these redirects.}
}
\value{
If t is a tibble, a new row is added and the tibble returned.
If t is NULL, a new tibble is created and returned.
}
\description{
Get tibbles from VIAF cluster using xml/xslt methods.

`cluster_summary` gets summary data, one row per viafID.

`cluster_sources` gets the sources with source ids.

`cluster_authorship`

`cluster_covers`

`cluster_ISBNs`

`cluster_jointcorps`

`cluster_jointpersons`

`cluster_links`

`cluster_names`

`cluster_titles`
}
\seealso{
\link{vc_summary}

\link{vc_sources}

\link{vc_authorship}

\link{vc_covers}

\link{vc_ISBNs}

\link{vc_jointcorps}

\link{vc_jointpersons}

\link{vc_viaflinks}

\link{vc_names}

\link{vc_titles}
}
