% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{viaf_get}
\alias{viaf_get}
\title{Get Data for VIAF Identifier(s)}
\usage{
viaf_get(query = NULL, ...)
}
\arguments{
\item{query}{The VIAF identifier(s) to get data for.}

\item{...}{Optional VIAF API query parameters.}
}
\value{
A tibble with data items.
}
\description{
Get authority cluster data based on supplied VIAF identifier(s).
}
\note{
An internet connection is required. The MARC 21 field
definitions are used.
}
\examples{
\donttest{viaf_get(c("64013650", "102333412"))}

}
