% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stylesheets.R
\name{get_stylesheet}
\alias{get_stylesheet}
\title{Get xslt stylesheet from package data}
\usage{
get_stylesheet(name = NULL)
}
\arguments{
\item{name}{The name of the stylesheet}
}
\value{
Loaded xslt stylesheet
}
\description{
Get an xslt stylesheet from the package data.
}
