% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stylesheets.R
\name{get_stylesheet_extdata}
\alias{get_stylesheet_extdata}
\title{Get xslt stylesheet from inst/extdata}
\usage{
get_stylesheet_extdata(name = NULL)
}
\arguments{
\item{name}{The name of the stylesheet.}
}
\value{
Loaded xslt stylesheet
}
\description{
Get an xslt stylesheet from the external data folder.
}
