% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{name_types}
\alias{name_types}
\title{VIAF Name Types}
\format{
A tibble with 5 rows and 2 variables.
}
\usage{
data(name_types)
}
\description{
A dataset containing the 5 possible name types in the VIAF.
}
\keyword{datasets}
