% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{viaf_search}
\alias{viaf_search}
\title{Search VIAF records}
\usage{
viaf_search(query = NULL, ...)
}
\arguments{
\item{query}{The search query (or queries) to get data for.}

\item{...}{Optional VIAF API query parameters.}
}
\value{
A named list of tibbles with data items.
}
\description{
Search VIAF records where the authority includes the given terms.
}
\note{
An internet connection is required. The MARC 21 field
definitions are used.
}
\examples{
\donttest{viaf_search(c("Rembrandt", "Jane Austen"))}

}
