% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggest.R
\name{viaf_suggest}
\alias{viaf_suggest}
\title{Suggest VIAF records}
\usage{
viaf_suggest(query = NULL, ...)
}
\arguments{
\item{query}{The search query (or queries) to get data for.}

\item{...}{Optional VIAF API query parameters.}
}
\value{
A named list of tibbles with data items.
}
\description{
Suggest VIAF records based on given terms passed in a query.
}
\note{
An internet connection is required.
}
\examples{
\donttest{viaf_suggest(c("rembrandt", "austen"))}

}
