% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stylesheets.R
\docType{data}
\name{vc_ISBNs_xsl}
\alias{vc_ISBNs_xsl}
\title{XSL stylesheet that extracts data from the ns1:ISBNs element of
VIAF cluster viaf.xml output}
\format{
An xsl stylesheet
}
\usage{
data(vc_ISBNs_xsl)
}
\description{
XSL stylesheet that extracts data from the ns1:ISBNs element of
VIAF cluster viaf.xml output
}
\keyword{xsl}
