% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stylesheets.R
\docType{data}
\name{vc_jointcorp_xsl}
\alias{vc_jointcorp_xsl}
\title{XSL stylesheet that extracts corporate joint author data from
the ns1:coauthors element of VIAF cluster viaf.xml output}
\format{
An xsl stylesheet
}
\usage{
data(vc_jointcorp_xsl)
}
\description{
XSL stylesheet that extracts corporate joint author data from
the ns1:coauthors element of VIAF cluster viaf.xml output
}
\keyword{xsl}
