% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stylesheets.R
\docType{data}
\name{vc_links_xsl}
\alias{vc_links_xsl}
\title{XSL stylesheet that extracts source link data from
the ns1:mainHeadings/ns1:mainHeadingEl elements of VIAF
cluster viaf.xml output}
\format{
An xsl stylesheet
}
\usage{
data(vc_links_xsl)
}
\description{
XSL stylesheet that extracts source link data from
the ns1:mainHeadings/ns1:mainHeadingEl elements of VIAF
cluster viaf.xml output
}
\keyword{xsl}
