% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stylesheets.R
\docType{data}
\name{vc_names_xsl}
\alias{vc_names_xsl}
\title{XSL stylesheet that extracts sources from
the ns1:mainHeadings/ns1:mainHeadingEl (preferred)
and ns1:x400s/ns1:x400 (alternate) elements of VIAF
cluster viaf.xml output}
\format{
An xsl stylesheet
}
\usage{
data(vc_names_xsl)
}
\description{
XSL stylesheet that extracts sources from
the ns1:mainHeadings/ns1:mainHeadingEl (preferred)
and ns1:x400s/ns1:x400 (alternate) elements of VIAF
cluster viaf.xml output
}
\keyword{xsl}
