% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xsl.R
\docType{data}
\name{vc_summary_xsl}
\alias{vc_summary_xsl}
\title{VIAF Cluster Summary xsl}
\format{
An xsl file
}
\usage{
data(vc_summary_xsl)
}
\description{
An XSL stylesheet that exports an xml data
file that summarizes a VIAF Cluster viaf.xml
data file
}
\keyword{viaf.xml}
\keyword{xslt}
