% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stylesheets.R
\docType{data}
\name{vc_titles_xsl}
\alias{vc_titles_xsl}
\title{XSL stylesheet that extracts name form data from
the ns1:titles/ns1:work elements of VIAF cluster viaf.xml output}
\format{
An xsl stylesheet
}
\usage{
data(vc_titles_xsl)
}
\description{
XSL stylesheet that extracts name form data from
the ns1:titles/ns1:work elements of VIAF cluster viaf.xml output
}
\keyword{xsl}
