% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterxml.R
\name{viaf_get_cluster}
\alias{viaf_get_cluster}
\title{Get VIAF cluster data in viaf.xml format}
\usage{
viaf_get_cluster(viafidlist = NULL, cache = tempdir(), ...)
}
\arguments{
\item{cache}{Path for storage of downloaded viaf.xml files. This is
to reduce redundant calls to the VIAF server. It also enables researchers to
maintain snapshot records of their raw data. The default is tempdir(), which
is deleted at the end of the current session.}

\item{viafIDlist}{The VIAF identifiers of clusters to be retrieved.}
}
\value{
A list of file paths to the downloaded viaf.xml files.
}
\description{
'viaf_get_cluster()' implements the VIAF API ''Get Data'' operation
}
\details{
Earlier versions of viafr (\url{https://github.com/stefanieschneider/viafr})
use json output from calls to the
\href{https://www.oclc.org/developer/api/oclc-apis/viaf/authority-cluster.en.html}{VIAF Authority Cluster Resource},
which provides data in 11 different formats including JSON, VIAF XML
and MARC21 XML. The
\href{https://www.oclc.org/developer/api/oclc-apis/viaf/authority-source.en.html}{Authority Source Resource},
on the other hand, only provides data in MARC21 and UNIMARC XML
formats. An objective of this extension to viafr is to enable access to
both cluster and source data.
VIAF authority clusters can be merged together over time
as additional data is received and algorithms are updated. Externally
maintained VIAF cluster identifiers should be checked for updates with
this call.
See
\href{https://www.oclc.org/developer/api/oclc-apis/viaf/authority-cluster.en.html}{Authority Cluster Resource}.
}
