% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterxml.R
\name{viaf_retrieve_xml}
\alias{viaf_retrieve_xml}
\title{Retrieve from api}
\usage{
viaf_retrieve_xml(endpoint = NULL, cache = tempdir(), ...)
}
\description{
Retrieve from api
}
\note{
Modified from rcrossref:::cr_GET and viaf_retrieve.
}
