% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viafxml2df.R
\name{viafxml2df}
\alias{viafxml2df}
\title{Data frame (tibble) from VIAF cluster xml}
\usage{
viafxml2df(x, xslt=NULL, ...)
}
\arguments{
\item{x}{Loaded viaf.xml document.}

\item{xslt}{Loaded xslt stylesheet}

\item{intercache}{Directory in which to story intermediate xml
output from transform.}
}
\value{
The target tibble.
}
\description{
'viafxml2df()' reads xml downloaded from viaf.org, transforms it to an
xml, and from that xml produces
a tibble. This function draws on Dominic Bordelon's
read_marcxml() function in
\href{https://github.com/dojobo/marc21r}{__marc21r__}.
}
