% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stylesheets.R, R/xsl.R
\docType{data}
\name{vc_summary_xsl}
\alias{vc_summary_xsl}
\title{XSL stylesheet making a summary of a VIAF cluster viaf.xml output}
\format{
An xsl stylesheet

An xsl file
}
\usage{
data(vc_summary_xsl)

data(vc_summary_xsl)
}
\description{
An XSL stylesheet that exports an xml data
file that summarizes a VIAF Cluster viaf.xml
data file
}
\keyword{viaf.xml}
\keyword{xsl}
\keyword{xslt}
